+++ CHANGELOG.md
@@ -38,6 +38,7 @@ CHANGELOG
 - `intelmq.bots.parsers.generic.parser_csv`:
   - New parameter `columns_required` to optionally ignore parse errors for columns.
 - added `intelmq.bots.parsers.cert_eu.parser_csv` (#1287).
+  - Do not overwrite the local `time.observation` with the data from the feed. The feed's field 'observation time' is now saved in the field `extra.cert_eu_time_observation`.
 - added `intelmq.bots.parsers.surbl.surbl`
 
 #### Experts